Consolidate into single component
